// ts-sdk/src/generated/types/marketHeader.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  combineCodec,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  getAddressDecoder,
  getAddressEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getStructDecoder,
  getStructEncoder,
  getU8Decoder,
  getU8Encoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
} from "@solana/kit";

export type MarketHeader = {
  discriminant: bigint;
  numSeats: number;
  numBids: number;
  numAsks: number;
  numFreeSectors: number;
  freeStackTop: number;
  seatsDllHead: number;
  seatsDllTail: number;
  bidsDllHead: number;
  bidsDllTail: number;
  asksDllHead: number;
  asksDllTail: number;
  baseMint: Address;
  quoteMint: Address;
  marketBump: number;
  numEvents: bigint;
  padding: Array<number>;
};

export type MarketHeaderArgs = {
  discriminant: number | bigint;
  numSeats: number;
  numBids: number;
  numAsks: number;
  numFreeSectors: number;
  freeStackTop: number;
  seatsDllHead: number;
  seatsDllTail: number;
  bidsDllHead: number;
  bidsDllTail: number;
  asksDllHead: number;
  asksDllTail: number;
  baseMint: Address;
  quoteMint: Address;
  marketBump: number;
  numEvents: number | bigint;
  padding: Array<number>;
};

export function getMarketHeaderEncoder(): FixedSizeEncoder<MarketHeaderArgs> {
  return getStructEncoder([
    ["discriminant", getU64Encoder()],
    ["numSeats", getU32Encoder()],
    ["numBids", getU32Encoder()],
    ["numAsks", getU32Encoder()],
    ["numFreeSectors", getU32Encoder()],
    ["freeStackTop", getU32Encoder()],
    ["seatsDllHead", getU32Encoder()],
    ["seatsDllTail", getU32Encoder()],
    ["bidsDllHead", getU32Encoder()],
    ["bidsDllTail", getU32Encoder()],
    ["asksDllHead", getU32Encoder()],
    ["asksDllTail", getU32Encoder()],
    ["baseMint", getAddressEncoder()],
    ["quoteMint", getAddressEncoder()],
    ["marketBump", getU8Encoder()],
    ["numEvents", getU64Encoder()],
    ["padding", getArrayEncoder(getU8Encoder(), { size: 3 })],
  ]);
}

export function getMarketHeaderDecoder(): FixedSizeDecoder<MarketHeader> {
  return getStructDecoder([
    ["discriminant", getU64Decoder()],
    ["numSeats", getU32Decoder()],
    ["numBids", getU32Decoder()],
    ["numAsks", getU32Decoder()],
    ["numFreeSectors", getU32Decoder()],
    ["freeStackTop", getU32Decoder()],
    ["seatsDllHead", getU32Decoder()],
    ["seatsDllTail", getU32Decoder()],
    ["bidsDllHead", getU32Decoder()],
    ["bidsDllTail", getU32Decoder()],
    ["asksDllHead", getU32Decoder()],
    ["asksDllTail", getU32Decoder()],
    ["baseMint", getAddressDecoder()],
    ["quoteMint", getAddressDecoder()],
    ["marketBump", getU8Decoder()],
    ["numEvents", getU64Decoder()],
    ["padding", getArrayDecoder(getU8Decoder(), { size: 3 })],
  ]);
}

export function getMarketHeaderCodec(): FixedSizeCodec<
  MarketHeaderArgs,
  MarketHeader
> {
  return combineCodec(getMarketHeaderEncoder(), getMarketHeaderDecoder());
}