// ts-sdk/src/generated/types/marketSeat.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  combineCodec,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  getAddressDecoder,
  getAddressEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
} from "@solana/kit";

export type MarketSeat = {
  user: Address;
  baseAvailable: bigint;
  quoteAvailable: bigint;
  userOrderSectors: {
    bids: Array<{ encodedPrice: number; sectorIndex: number }>;
    asks: Array<{ encodedPrice: number; sectorIndex: number }>;
  };
};

export type MarketSeatArgs = {
  user: Address;
  baseAvailable: number | bigint;
  quoteAvailable: number | bigint;
  userOrderSectors: {
    bids: Array<{ encodedPrice: number; sectorIndex: number }>;
    asks: Array<{ encodedPrice: number; sectorIndex: number }>;
  };
};

export function getMarketSeatEncoder(): FixedSizeEncoder<MarketSeatArgs> {
  return getStructEncoder([
    ["user", getAddressEncoder()],
    ["baseAvailable", getU64Encoder()],
    ["quoteAvailable", getU64Encoder()],
    [
      "userOrderSectors",
      getStructEncoder([
        [
          "bids",
          getArrayEncoder(
            getStructEncoder([
              ["encodedPrice", getU32Encoder()],
              ["sectorIndex", getU32Encoder()],
            ]),
            { size: 10 },
          ),
        ],
        [
          "asks",
          getArrayEncoder(
            getStructEncoder([
              ["encodedPrice", getU32Encoder()],
              ["sectorIndex", getU32Encoder()],
            ]),
            { size: 10 },
          ),
        ],
      ]),
    ],
  ]);
}

export function getMarketSeatDecoder(): FixedSizeDecoder<MarketSeat> {
  return getStructDecoder([
    ["user", getAddressDecoder()],
    ["baseAvailable", getU64Decoder()],
    ["quoteAvailable", getU64Decoder()],
    [
      "userOrderSectors",
      getStructDecoder([
        [
          "bids",
          getArrayDecoder(
            getStructDecoder([
              ["encodedPrice", getU32Decoder()],
              ["sectorIndex", getU32Decoder()],
            ]),
            { size: 10 },
          ),
        ],
        [
          "asks",
          getArrayDecoder(
            getStructDecoder([
              ["encodedPrice", getU32Decoder()],
              ["sectorIndex", getU32Decoder()],
            ]),
            { size: 10 },
          ),
        ],
      ]),
    ],
  ]);
}

export function getMarketSeatCodec(): FixedSizeCodec<
  MarketSeatArgs,
  MarketSeat
> {
  return combineCodec(getMarketSeatEncoder(), getMarketSeatDecoder());
}