// ts-sdk/src/generated/types/order.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getStructDecoder,
  getStructEncoder,
  getU8Decoder,
  getU8Encoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
} from "@solana/kit";

export type Order = {
  encodedPrice: number;
  userSeatIndex: number;
  baseRemaining: bigint;
  quoteRemaining: bigint;
  padding: Array<number>;
};

export type OrderArgs = {
  encodedPrice: number;
  userSeatIndex: number;
  baseRemaining: number | bigint;
  quoteRemaining: number | bigint;
  padding: Array<number>;
};

export function getOrderEncoder(): FixedSizeEncoder<OrderArgs> {
  return getStructEncoder([
    ["encodedPrice", getU32Encoder()],
    ["userSeatIndex", getU32Encoder()],
    ["baseRemaining", getU64Encoder()],
    ["quoteRemaining", getU64Encoder()],
    ["padding", getArrayEncoder(getU8Encoder(), { size: 184 })],
  ]);
}

export function getOrderDecoder(): FixedSizeDecoder<Order> {
  return getStructDecoder([
    ["encodedPrice", getU32Decoder()],
    ["userSeatIndex", getU32Decoder()],
    ["baseRemaining", getU64Decoder()],
    ["quoteRemaining", getU64Decoder()],
    ["padding", getArrayDecoder(getU8Decoder(), { size: 184 })],
  ]);
}

export function getOrderCodec(): FixedSizeCodec<OrderArgs, Order> {
  return combineCodec(getOrderEncoder(), getOrderDecoder());
}