// ts-sdk/src/generated/types/sector.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getStructDecoder,
  getStructEncoder,
  getU8Decoder,
  getU8Encoder,
  getU32Decoder,
  getU32Encoder,
} from "@solana/kit";

export type Sector = { next: number; prev: number; payload: Array<number> };

export type SectorArgs = Sector;

export function getSectorEncoder(): FixedSizeEncoder<SectorArgs> {
  return getStructEncoder([
    ["next", getU32Encoder()],
    ["prev", getU32Encoder()],
    ["payload", getArrayEncoder(getU8Encoder(), { size: 208 })],
  ]);
}

export function getSectorDecoder(): FixedSizeDecoder<Sector> {
  return getStructDecoder([
    ["next", getU32Decoder()],
    ["prev", getU32Decoder()],
    ["payload", getArrayDecoder(getU8Decoder(), { size: 208 })],
  ]);
}

export function getSectorCodec(): FixedSizeCodec<SectorArgs, Sector> {
  return combineCodec(getSectorEncoder(), getSectorDecoder());
}